Solve ImportDialog issues
